Implement the task with C++ stdlib and Windows SDK
This commit is another implementation of the test task.  The first
solution was made in 1132b4b (Implement the task with C++ standard
library, 2021-08-23).  This commit still uses C++ standard library (in
the end of the day -- this is a C++ project!), but it replaces:

  * std::thread with Microsoft Windows native thread pool (one from
    Microsoft Windows Vista, not XP).
  * std::mutex with Microsoft Windows ``Slim reader/writer lock''.
  * std::condition_variable with Microsoft Windows condition variable
    native implementation.

Therefore, this is not a pure Microsoft Windows SDK/API solution.  But
it replaced some major parts of the previous one.

Signed-off-by: Ruslan Garipov <ruslanngaripov@gmail.com>