Implement the task with C and "plain" threads
This is the last (I believe) implementation of the test task.  This one
uses C (and its stdlib) and Windows SDK.  It is very similar to
implementation in 7676095 (Implement the task with C and Windows SDK,
2021-08-31), but this commit uses Windows SDK's "raw" threads created
with _beginthreadex().

See also: 1132b4b (Implement the task with C++ standard library,
2021-08-23), 9f4effe (Implement the task with C++ stdlib and Windows
SDK, 2021-08-26) and 7676095 (Implement the task with C and Windows SDK,
2021-08-31)

Signed-off-by: Ruslan Garipov <ruslanngaripov@gmail.com>